move headers to 'include' folder
To be consistent to the new sensors structure, public headers have been moved to
the include folder. Those header may still contain some private function. Those
files can be split in two headers (public and private) in future
