+++ API/sensors/utility.c
@@ -1,5 +1,5 @@
 #include "../inputs/ev3_input_uart.h"
-#include "../ev3_command.h"
+#include "ev3_command.h"
 #include "utility.h"
 
 void setUARTSensorHandlerMode(SensorHandler *handler, int port, int type, int mode) {
